# Performance improvements to _LegacyRecordBatchBuilderPy