--- a/aiokafka/record/util.py
+++ b/aiokafka/record/util.py
@@ -1,8 +1,0 @@
-import binascii
-
-
-def calc_crc32(memview):
-    """ Calculate simple CRC-32 checksum over a memoryview of data
-    """
-    crc = binascii.crc32(memview)
-    return crc