Highlights include:
- Perform `append()` type checking only when necessary (upon error)
- Wrap all `_encode_msg()` logic into a single `struct.pack_into()` call to reduce overhead
- Flatten multi-method invocations wherever possible without losing clarity
- Preallocate buffer space rather than expanding on each individual message
- Calculate key and value length only once

The results:
```
(.venv) aiokafka (master)$ python benchmark/record_batch_compose.py
.....................
batch_append_v0: Mean +- std dev: 906 us +- 60 us
.....................
batch_append_v1: Mean +- std dev: 946 us +- 52 us
(.venv) aiokafka (master)$ python bench_records.py simple
1000000 messages in 7.83s (127758.34 msg/s)

(.venv) aiokafka (shargan/record-performance)$ python benchmark/record_batch_compose.py
.....................
batch_append_v0: Mean +- std dev: 598 us +- 29 us
.....................
batch_append_v1: Mean +- std dev: 647 us +- 35 us
(.venv) aiokafka (shargan/record-performance)$ python bench_records.py simple
1000000 messages in 5.50s (181839.19 msg/s)
```