@shargan 2017-10-25
bench_records.py is a stupid-simple micro-benchmark I was using mostly for profiling:
```python
import random
import sys
import time

from aiokafka.producer.message_accumulator import BatchBuilder

BATCH_SIZE = 1024 * 128
MAGIC = 1

def bench_simple(num, payload):
    magic = MAGIC
    size = BATCH_SIZE

    batch = BatchBuilder(MAGIC, BATCH_SIZE, None)
    i = 0
    while i < num:
        if batch.append(timestamp=None, key=None, value=payload) is None:
            batch._build()
            batch = BatchBuilder(MAGIC, BATCH_SIZE, None)
        else:
            i += 1

def main(arg):
    funcs = {"simple": bench_simple}
    num = 1000000
    size = 100
    payload = ("m" * size).encode()

    bench = funcs[arg]
    start = time.time()
    bench(num, payload)
    delta = time.time() - start
    print("%d messages in %.2fs (%.2f msg/s)" % (num, delta, num / delta))

if __name__ == "__main__":
    main(sys.argv[1])
```

Profiling master vs. these changes shows the substance of these gains. Here's the most interesting values from each:
```
VERSION    FUNC              CALLS      TOTTIME     CUMTIME
master     append()          1001022    3.145       10.812
PR         append()          1001022    2.204       6.164

master     _encode_msg()     1000000    3.696       5.400
PR         _encode_msg()     1000000    2.192       3.127

master     size_in_bytes()   1001022    0.552       1.449
PR         size_in_bytes()   1001022    0.342       0.342

master     calc_crc32()      1000000    0.349       0.606
PR         binascii.crc32()  1000000    0.241       0.241
```