@shargan 2017-10-27
Can you clarify what you find harder to read in this code? Personally, I find it quite a bit easier to read but that's a style-preference thing. The diff makes it look a lot uglier than it is. Is there something in particular you find hard to parse?