@shargan 2017-11-08
The most recent diff changes some behavior as requested. Does extending the buffer in chunks work for you as a middle ground? We still get most of the performance benefit while giving up a little bit to repeated `len()` calls.

```bash
# fully pre-allocated buffer, no len() calls
$ python benchmark/record_batch_compose.py
.....................
batch_append_v0: Mean +- std dev: 616 us +- 43 us
.....................
batch_append_v1: Mean +- std dev: 667 us +- 33 us
$ python bench_records.py simple
1000000 messages in 5.44s (183712.12 msg/s)


# chunk-allocated buffer with len() calls
$ python benchmark/record_batch_compose.py
.....................
batch_append_v0: Mean +- std dev: 602 us +- 42 us
.....................
batch_append_v1: Mean +- std dev: 648 us +- 39 us
$ python bench_records.py simple
1000000 messages in 5.64s (177179.21 msg/s)


# no pre-allocation, extended by message size on each call
$ python benchmark/record_batch_compose.py
.....................
batch_append_v0: Mean +- std dev: 600 us +- 43 us
.....................
batch_append_v1: Mean +- std dev: 623 us +- 19 us
$ python bench_records.py simple
1000000 messages in 6.34s (157789.73 msg/s)
```

Note that benchmark/record_batch_compose.py shows no positive change from this because it only adds 100 messages per batch. The performance improvement here really becomes apparent as the batches are fully utilized and the cost of extending them surpasses the cost of allocation. To compare, here's the results of that benchmark with `MESSAGES_PER_BATCH=2000`:

```bash
# no pre-allocation, extended by message size on each call
$ python benchmark/record_batch_compose.py
.....................
batch_append_v0: Mean +- std dev: 11.3 ms +- 0.5 ms
.....................
batch_append_v1: Mean +- std dev: 13.3 ms +- 0.9 ms


# chunk-allocated buffer with len() calls
$ python benchmark/record_batch_compose.py
.....................
batch_append_v0: Mean +- std dev: 10.7 ms +- 0.3 ms
.....................
batch_append_v1: Mean +- std dev: 12.0 ms +- 0.8 ms
```