@tvoinarovskyi 2018-01-25
Hey there, I would like to have this PR merge to release, @shargan If you can/like to apply changes, please say so. If not it's not a problem, will merge like this and make minor tweeks in a followup. Sorry for being too picky with those changes and the ones before, I am very grateful for your work on this)