@shargan 2018-02-01
Hey @tvoinarovskyi, sorry for dropping off the radar. Deadlines have prevented me from making progress on this and will likely continue to do so for a couple more weeks. I don't mind the pickiness; I'll be sure to send much smaller PRs in the future to make changes easier to work through. Thanks for working with me on this.