@tvoinarovskyi 2018-02-01
@shargan No problem, I did drop off lately myself) Did a followup on this https://github.com/aio-libs/aiokafka/pull/301